minor fixes following reviewers' suggestions
